refactor(trace): split standard probe helpers
